backend/requirements.txt
flask==2.3.2
# Lighter transformer version
transformers==4.28.0
# CPU-only PyTorch
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1
torchvision==0.15.2
# Minimal OCR dependencies
easyocr==1.7.2
python-dotenv==1.0.0
gunicorn==20.1.0
# Add explicit dependencies for smaller models
sentencepiece  # Required for DistilBERT
protobuf<=3.20.0  # Compatibility with older transformers
numpy